Bug fixes and connected loaded torrents to store
Fixed a particularly insidious bug with the data store being shared
across the worker threads. Instantiating the shared object inside of the
HttpServer declaration actually led to each worker getting its own local
state instead of the shared state between all of them. This led to an
incorrect return of all swarm, peer, and torrent information as each
thread could have completely different values in their stores. Moving
the instantiation outside of the HttpServer declaration (more
specifically, from inside the move closure to just before the
declaration) completely fixed the problem.

Fixed a bug in which the decrementing operation to the incomplete
metric of a torrent would lead to a panic if that metric became
negative.

Torrent records loaded from MySQL are now loaded into the data store
object and are able to be interacted upon through requests from clients.