Implemented client approval middleware
This is the first step in private tracker support. A list of clients can
be defined in config.yaml by using their corresponding peer IDs along
with their versions. The middleware will then load the list as well as other
parameters and on every request, it will check to see if the client is
either allowed or prohibited (depending on configuration) and pass it to
the parsing step or reject the request with a message for the client,
respectively.

Also fixed a small bug in which the peer key was not properly parsed. So maybe
that's a big bug, given that it's pretty important. Oh well, it's fixed
now.