Added a CompactPeer object
A CompactPeer object has been added in order to reduce the time spent in
get_peers(). The tracker returns compact responses by default, so peer
IDs are not sent in tracker responses. The get_peers() function had to
clone peers out of the hashsets in order to eventually pass them to the
AnnounceResponse function. So, by using the full peer information to
instantiate the compact peers, the latter can be used for the tracker
response and remove cloning (and excessive heap allocation) from
get_peers() entirely.